Include CMU Serif, simplify CSS imports (#72)
This uses my new [`computer-modern`
package](https://npmjs.com/package/computer-modern) to make Computer
Modern Serif available directly from Mafs.

This changes the CSS import in Mafs:

```diff
-@import "mafs/build/index.css";
+@import "mafs/core.css";
+@import "mafs/font.css";
```

`font.css` is optional. It just imports `computer-modern/cmun-serif.css`
and sets the MafsView font.

---

One thing that might be handy in the future would be to subset the fonts
to only have numbers and common characters, so that the browser doesn't
have to load so much content just to show axes.