Add basic support for polylines (i.e. open polygons) (#112)
As the title suggests, this PR adds support for the `<polyline>` SVG
element. Since this element is so similar to the `<polygon>` SVG
element, I figured it'd be better to add a prop to switch between the
two, instead of adding a separate `<Polyline>` Mafs component.

I've only really added the prop and an example so far and have not
updated any docs/tests yet. Let me know if you agree with this approach
and I'll get on those docs and tests.

---------

Co-authored-by: Steven Petryk <petryk.steven@gmail.com>