Better handling of canvas width (#114)
This PR implements both suggested solutions from #113:
- The "canvas" is only rendered once we have an actual width. To achieve
this I've lifted part of the `<Mafs>` component into an internal
`<MafsCanvas>` component. Now all the hooks and render logic in
`<MafsCanvas>` can safely assume the width is non-zero.
- If the user passes an explicit pixel value (i.e. not "auto") to the
`width` prop, that value will be used directly and the
`useResizeObserver` is disabled. To disable the resize observer, I used
the method suggested here:
https://github.com/ZeeCoder/use-resize-observer#opting-out-of-or-delaying-resizeobserver-instantiation.

The code requires some cleaning up still:
- [ ] Passing each prop from `<Mafs>` to `<MafsCanvas>` individually is
annoying, but I think it is necessary in order for the docgen to pick up
the default values. But maybe there is a better way?
- [ ] There is a `visible` state that is enabled after the component is
mounted, triggering a re-render to actually show the contents. I think
this might not be necessary anymore, since we now always ensure the
canvas has a size?

Fixes #113

---------

Co-authored-by: Steven Petryk <petryk.steven@gmail.com>